requirements.txt
numpy==1.23.4
Pillow==9.3.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.14.0+cpu
tqdm==4.64.1
python-dotenv==1.0.0
wget==3.2
